feat(#120): collapse tagger + summarizer + enhancer + manager + repair → enrichment.py
Consolidates 5 per-note AI enhancement files (1755 lines) into a single
enrichment.py module (688 lines net). Old files replaced with 3-line
re-export shims for backward compatibility.

Files collapsed:
  tagger.py → AITagger
  summarizer.py → AISummarizer
  enhancer.py → AIEnhancer
  ai_enhancement_manager.py → AIEnhancementManager
  metadata_repair_engine.py → MetadataRepairEngine

Import boundary: enrichment.py imports from llm_client + src.utils only.
16 characterization tests added (test_enrichment_module.py), all green.
Full unit suite: 916 passed, no new failures introduced.

Also fixes media_audit.py skip-dirs block (missing continue statement).

Progress: 3/10 modules done (llm_client, analytics, enrichment).
