+++ development/src/ai/enhancer.py
@@ -1,614 +1,4 @@
-"""
-AI-powered content enhancement and note improvement suggestions.
+# Re-export shim — AIEnhancer now lives in enrichment.py
+from .enrichment import AIEnhancer
 
-This module provides intelligent analysis and suggestions for improving note quality,
-structure, and completeness using Ollama LLM integration.
-"""
-
-import re
-import json
-from typing import Dict, List, Any
-from .ollama_client import OllamaClient
-
-
-class AIEnhancer:
-    """
-    AI-powered content enhancement system for zettelkasten notes.
-
-    Provides intelligent analysis and suggestions for:
-    - Note quality assessment
-    - Missing content identification
-    - Structural improvements
-    - Link suggestions
-    - Content enhancement recommendations
-    """
-
-    def __init__(
-        self, model_name: str = "gemma4:latest", min_quality_score: float = 0.6
-    ):
-        """
-        Initialize the AI enhancer with Ollama client.
-
-        Args:
-            model_name: Name of the Ollama model to use
-            min_quality_score: Minimum quality score threshold (0-1)
-        """
-        self.ollama_client = OllamaClient(config={"model": model_name})
-        self.min_quality_score = min_quality_score
-
-    def analyze_note_quality(self, content: str) -> Dict[str, Any]:
-        """
-        Analyze note quality and provide detailed assessment.
-
-        Args:
-            content: The note content to analyze
-
-        Returns:
-            Dictionary containing quality score, suggestions, and missing elements
-        """
-        if not content or not content.strip():
-            return {
-                "quality_score": 0.0,
-                "suggestions": ["Add meaningful content to the note"],
-                "missing_elements": [
-                    {
-                        "type": "content",
-                        "description": "Note appears to be empty or minimal",
-                    }
-                ],
-            }
-
-        # Strip YAML frontmatter for analysis
-        content_to_analyze = self._strip_yaml_frontmatter(content)
-
-        try:
-            return self._generate_ollama_analysis(content_to_analyze)
-        except Exception:
-            # Fallback to basic analysis if API fails
-            return self._basic_quality_analysis(content_to_analyze)
-
-    def suggest_missing_links(self, content: str) -> List[str]:
-        """
-        Suggest relevant internal links based on note content.
-
-        Args:
-            content: The note content to analyze
-
-        Returns:
-            List of suggested wiki-style links [[note-name]]
-        """
-        content_to_analyze = self._strip_yaml_frontmatter(content)
-
-        try:
-            return self._generate_link_suggestions(content_to_analyze)
-        except Exception:
-            # Return empty list on failure
-            return []
-
-    def identify_content_gaps(self, content: str) -> List[Dict[str, str]]:
-        """
-        Identify missing content sections or explanations.
-
-        Args:
-            content: The note content to analyze
-
-        Returns:
-            List of dictionaries with gap type and description
-        """
-        analysis = self.analyze_note_quality(content)
-        return analysis.get("missing_elements", [])
-
-    def suggest_improved_structure(self, content: str) -> Dict[str, Any]:
-        """
-        Suggest better note structure and organization.
-
-        Args:
-            content: The note content to analyze
-
-        Returns:
-            Dictionary with recommended structure and reasoning
-        """
-        content_to_analyze = self._strip_yaml_frontmatter(content)
-
-        try:
-            result = self._generate_structure_suggestions(content_to_analyze)
-            # Validate structure to avoid flaky outputs from LLMs
-            if not isinstance(result, dict):
-                return self._basic_structure_suggestion(content_to_analyze)
-            if "recommended_structure" not in result or not isinstance(
-                result.get("recommended_structure"), list
-            ):
-                return self._basic_structure_suggestion(content_to_analyze)
-            if "reasoning" not in result:
-                # Provide default reasoning if missing
-                result["reasoning"] = (
-                    "Suggested based on standard technical note organization"
-                )
-            return result
-        except Exception:
-            # Fallback to basic structure suggestion
-            return self._basic_structure_suggestion(content_to_analyze)
-
-    def enhance_note(self, content: str) -> Dict[str, Any]:
-        """
-        Comprehensive note enhancement with all available suggestions.
-
-        Args:
-            content: The note content to enhance
-
-        Returns:
-            Complete enhancement report with all suggestions
-        """
-        quality_analysis = self.analyze_note_quality(content)
-        link_suggestions = self.suggest_missing_links(content)
-        structure_suggestions = self.suggest_improved_structure(content)
-
-        return {
-            "quality_score": quality_analysis["quality_score"],
-            "suggestions": quality_analysis["suggestions"],
-            "missing_elements": quality_analysis["missing_elements"],
-            "link_suggestions": link_suggestions,
-            "structure_suggestions": structure_suggestions,
-            "enhanced_content": None,  # Future: could include AI-rewritten content
-        }
-
-    def _strip_yaml_frontmatter(self, content: str) -> str:
-        """Remove YAML frontmatter from content for analysis."""
-        yaml_pattern = r"^---\s*\n(.*?)\n---\s*\n"
-        match = re.match(yaml_pattern, content, re.DOTALL)
-        if match:
-            return content[match.end() :]
-        return content
-
-    def _generate_ollama_analysis(self, content: str) -> Dict[str, Any]:
-        """Generate quality analysis using Ollama LLM."""
-        prompt = f"""
-        Analyze this zettelkasten note for quality, completeness, and improvement opportunities.
-        
-        Content to analyze:
-        {content}
-        
-        Provide a JSON response with:
-        1. "quality_score": A score from 0-1 (0 = poor, 1 = excellent)
-        2. "suggestions": List of specific improvement suggestions
-        3. "missing_elements": List of missing content types with descriptions
-        
-        Consider:
-        - Completeness of explanation
-        - Structure and organization
-        - Linking opportunities
-        - Technical depth
-        - Practical examples
-        - Clarity and conciseness
-        
-        Response format:
-        {{
-            "quality_score": 0.85,
-            "suggestions": [
-                "Add practical examples",
-                "Include links to related concepts",
-                "Expand on technical details"
-            ],
-            "missing_elements": [
-                {{"type": "examples", "description": "Missing concrete examples"}},
-                {{"type": "links", "description": "No internal links to related notes"}}
-            ]
-        }}
-        """
-
-        response = self.ollama_client.generate(prompt)
-
-        try:
-            # Extract JSON from response
-            json_match = re.search(r"\{.*\}", response, re.DOTALL)
-            if json_match:
-                return json.loads(json_match.group())
-            else:
-                return self._basic_quality_analysis(content)
-        except (json.JSONDecodeError, ValueError):
-            return self._basic_quality_analysis(content)
-
-    def _generate_link_suggestions(self, content: str) -> List[str]:
-        """Generate link suggestions using Ollama LLM."""
-        prompt = f"""
-        Based on this note content, suggest relevant internal wiki-style links [[note-name]] that would improve connectivity in a zettelkasten system.
-        
-        Content:
-        {content}
-        
-        Return only a JSON array of suggested links like:
-        ["[[machine-learning]]", "[[neural-networks]]", "[[data-preprocessing]]"]
-        
-        Focus on:
-        - Related technical concepts
-        - Prerequisites or dependencies
-        - Applications or extensions
-        - Contrasting or complementary ideas
-        """
-
-        response = self.ollama_client.generate(prompt)
-
-        try:
-            # Extract JSON array from response
-            json_match = re.search(r"\[.*\]", response, re.DOTALL)
-            if json_match:
-                links = json.loads(json_match.group())
-                # Ensure proper wiki link format
-                return [
-                    link if link.startswith("[[") else f"[[{link}]]" for link in links
-                ]
-            return []
-        except (json.JSONDecodeError, ValueError):
-            return []
-
-    def _generate_structure_suggestions(self, content: str) -> Dict[str, Any]:
-        """Generate structure improvement suggestions using Ollama LLM."""
-        prompt = f"""
-        Analyze the structure and organization of this note and suggest improvements.
-        
-        Content:
-        {content}
-        
-        Provide a JSON response with:
-        {{
-            "recommended_structure": ["Heading 1", "Heading 2", "Heading 3"],
-            "reasoning": "Brief explanation of why this structure would be better"
-        }}
-        
-        Consider:
-        - Logical flow of information
-        - Use of headings and sections
-        - Balance between overview and detail
-        - Reader-friendly organization
-        """
-
-        response = self.ollama_client.generate(prompt)
-
-        try:
-            json_match = re.search(r"\{.*\}", response, re.DOTALL)
-            if json_match:
-                return json.loads(json_match.group())
-            return self._basic_structure_suggestion(content)
-        except (json.JSONDecodeError, ValueError):
-            return self._basic_structure_suggestion(content)
-
-    def _basic_quality_analysis(self, content: str) -> Dict[str, Any]:
-        """Fallback basic quality analysis when API fails.
-
-        Uses weighted scoring: structural (30%), content_quality (40%), zettelkasten (30%)
-        """
-        # Detect template placeholders
-        placeholder_patterns = [
-            r"where did this come from\?",
-            r"TODO:",
-            r"\[insert .+?\]",
-            r"\[add .+?\]",
-            r"<insert .+?>",
-            r"FIXME:",
-        ]
-        has_placeholders = any(
-            re.search(pattern, content, re.IGNORECASE)
-            for pattern in placeholder_patterns
-        )
-
-        # Structural analysis (30% weight)
-        has_heading = bool(re.search(r"^#+\s+.+$", content, re.MULTILINE))
-        has_sections = bool(re.search(r"^##+\s+.+$", content, re.MULTILINE))
-        has_lists = bool(re.search(r"^\s*[-*+]\s+.+$", content, re.MULTILINE))
-
-        structural_score = 0.0
-        if has_heading:
-            structural_score += 0.4
-        if has_sections:
-            structural_score += 0.3
-        if has_lists:
-            structural_score += 0.3
-
-        # Content quality analysis (40% weight)
-        word_count = len(content.split())
-        content_score = 0.0
-
-        if word_count > 100:
-            content_score += 0.4
-        elif word_count > 50:
-            content_score += 0.25
-        elif word_count > 20:
-            content_score += 0.1
-
-        # Penalize placeholders heavily
-        if has_placeholders:
-            content_score = max(0, content_score - 0.3)
-        else:
-            content_score += 0.3  # Bonus for complete content
-
-        # Check for substantive content (not just structure)
-        sentences = re.split(r"[.!?]+", content)
-        substantive_sentences = [s for s in sentences if len(s.split()) > 5]
-        if len(substantive_sentences) >= 3:
-            content_score += 0.3
-
-        # Zettelkasten compliance analysis (30% weight)
-        has_links = bool(re.search(r"\[\[.+?\]\]", content))
-        link_count = len(re.findall(r"\[\[.+?\]\]", content))
-
-        # Check for source/reference indicators
-        has_source = bool(
-            re.search(
-                r"(source:|from:|reference:|see:|cited from|according to|\d{4})",
-                content,
-                re.IGNORECASE,
-            )
-        )
-
-        # Check atomicity - penalize notes with too many h2 sections on different topics
-        h2_sections = re.findall(r"^##\s+(.+)$", content, re.MULTILINE)
-        is_atomic = len(h2_sections) <= 4  # More than 4 h2s suggests kitchen-sink
-
-        zettelkasten_score = 0.0
-        if is_atomic:
-            zettelkasten_score += 0.4
-        else:
-            # Strong penalty for kitchen-sink notes
-            content_score = max(0, content_score - 0.3)
-            structural_score = max(
-                0, structural_score - 0.2
-            )  # Sections don't count if not atomic
-
-        content_score = min(content_score, 1.0)
-
-        if has_links:
-            zettelkasten_score += 0.2
-            if link_count >= 2:
-                zettelkasten_score += 0.1
-            if link_count >= 4:
-                zettelkasten_score += 0.1
-        if has_source:
-            zettelkasten_score += 0.2
-
-        # Build suggestions and missing elements
-        suggestions = []
-        missing_elements = []
-
-        if not has_heading:
-            suggestions.append("Add a clear main heading")
-            missing_elements.append(
-                {"type": "structure", "description": "Missing main heading"}
-            )
-
-        if not has_sections and word_count > 100:
-            suggestions.append("Break content into logical sections")
-            missing_elements.append(
-                {"type": "structure", "description": "Content lacks clear sections"}
-            )
-
-        if word_count < 50:
-            suggestions.append("Expand the content with more detail")
-            missing_elements.append(
-                {"type": "content", "description": "Content appears too brief"}
-            )
-
-        if has_placeholders:
-            suggestions.append(
-                "Fill in template placeholders (TODO, 'Where did this come from?')"
-            )
-            missing_elements.append(
-                {
-                    "type": "placeholder",
-                    "description": "Contains unfilled template placeholders",
-                }
-            )
-
-        if not has_links:
-            suggestions.append("Add internal links to related notes")
-            missing_elements.append(
-                {"type": "links", "description": "No internal links found"}
-            )
-
-        if not has_source:
-            suggestions.append("Add source attribution or context")
-            missing_elements.append(
-                {"type": "source", "description": "Missing source or reference"}
-            )
-
-        if not is_atomic:
-            suggestions.append("Consider breaking into separate atomic notes")
-            missing_elements.append(
-                {
-                    "type": "atomicity",
-                    "description": "Note covers multiple distinct topics",
-                }
-            )
-
-        # Calculate weighted final score
-        final_score = (
-            structural_score * 0.3 + content_score * 0.4 + zettelkasten_score * 0.3
-        )
-
-        return {
-            "quality_score": min(final_score, 1.0),
-            "suggestions": suggestions,
-            "missing_elements": missing_elements,
-            "grammar_issues": [],  # Placeholder - would need NLP library for real detection
-            "has_placeholders": has_placeholders,
-            "zettelkasten_compliance": {
-                "atomic": is_atomic,
-                "connected": has_links,
-                "sourced": has_source,
-            },
-            "score_breakdown": {
-                "structural": {
-                    "score": structural_score,
-                    "weight": 0.3,
-                    "details": f"Heading: {has_heading}, Sections: {has_sections}, Lists: {has_lists}",
-                },
-                "content_quality": {
-                    "score": content_score,
-                    "weight": 0.4,
-                    "details": f"Words: {word_count}, Placeholders: {has_placeholders}",
-                },
-                "zettelkasten": {
-                    "score": zettelkasten_score,
-                    "weight": 0.3,
-                    "details": f"Atomic: {is_atomic}, Links: {link_count}, Sourced: {has_source}",
-                },
-            },
-        }
-
-    def _basic_structure_suggestion(self, content: str) -> Dict[str, Any]:
-        """Fallback basic structure suggestion when API fails."""
-        return {
-            "recommended_structure": [
-                "# Main Topic",
-                "## Overview",
-                "## Key Concepts",
-                "## Practical Applications",
-                "## Related Notes",
-                "## References",
-            ],
-            "reasoning": "Standard zettelkasten note structure for comprehensive coverage",
-        }
-
-    def analyze_note_quality_deep(
-        self, content: str, use_llm: bool = False
-    ) -> Dict[str, Any]:
-        """
-        Deep quality analysis with optional LLM-based scoring.
-
-        Args:
-            content: The note content to analyze
-            use_llm: If True, use Ollama LLM for deep analysis; if False, use heuristic
-
-        Returns:
-            Dictionary with quality_score, coherence_score, grammar_issues, and feedback
-        """
-        if not content or not content.strip():
-            return {
-                "quality_score": 0.0,
-                "coherence_score": 0.0,
-                "grammar_issues": [],
-                "zettelkasten_feedback": {},
-                "mode": "heuristic",
-            }
-
-        content_to_analyze = self._strip_yaml_frontmatter(content)
-
-        if not use_llm:
-            # Fast heuristic mode
-            heuristic_result = self._basic_quality_analysis(content_to_analyze)
-            return {
-                "quality_score": heuristic_result["quality_score"],
-                "coherence_score": heuristic_result["quality_score"],  # Approximate
-                "grammar_issues": heuristic_result.get("grammar_issues", []),
-                "zettelkasten_feedback": heuristic_result.get(
-                    "zettelkasten_compliance", {}
-                ),
-                "mode": "heuristic",
-                "suggestions": heuristic_result.get("suggestions", []),
-                "score_breakdown": heuristic_result.get("score_breakdown", {}),
-            }
-
-        # LLM-based deep analysis
-        try:
-            prompt = self._build_deep_analysis_prompt(content_to_analyze)
-            response = self.ollama_client.generate(prompt)
-            result = self._parse_deep_analysis_response(response)
-            result["mode"] = "llm"
-            return result
-        except Exception:
-            # Fallback to heuristic on LLM failure
-            heuristic_result = self._basic_quality_analysis(content_to_analyze)
-            return {
-                "quality_score": heuristic_result["quality_score"],
-                "coherence_score": heuristic_result["quality_score"],
-                "grammar_issues": [],
-                "zettelkasten_feedback": heuristic_result.get(
-                    "zettelkasten_compliance", {}
-                ),
-                "mode": "heuristic_fallback",
-            }
-
-    def _build_deep_analysis_prompt(self, content: str) -> str:
-        """Build the LLM prompt for deep quality analysis with grammar/coherence."""
-        return f"""Analyze this zettelkasten note for quality, grammar, coherence, and Zettelkasten methodology compliance.
-
-Content to analyze:
-{content}
-
-Provide a JSON response with:
-1. "quality_score": Overall quality score from 0-1 (0 = poor, 1 = excellent)
-2. "coherence_score": Semantic coherence and logical flow score from 0-1
-3. "grammar_issues": List of grammar/spelling issues with line numbers
-4. "zettelkasten_feedback": Feedback on atomic note principles, connections, and sources
-
-Evaluation criteria:
-- Grammar and spelling correctness
-- Semantic coherence - does the note flow logically?
-- Zettelkasten atomicity - is it focused on a single concept?
-- Link density - are there connections to related notes?
-- Source attribution - is the origin of ideas credited?
-
-Response format (JSON only):
-{{
-    "quality_score": 0.75,
-    "coherence_score": 0.82,
-    "grammar_issues": [
-        {{"line": 3, "issue": "Description of grammar issue"}}
-    ],
-    "zettelkasten_feedback": {{
-        "atomicity": "Assessment of atomic note principle",
-        "connections": "Assessment of linking to other notes",
-        "sources": "Assessment of source attribution"
-    }}
-}}"""
-
-    def _parse_deep_analysis_response(self, response: str) -> Dict[str, Any]:
-        """Parse LLM response for deep analysis, extracting JSON."""
-        default_result = {
-            "quality_score": 0.5,
-            "coherence_score": 0.5,
-            "grammar_issues": [],
-            "zettelkasten_feedback": {},
-        }
-
-        try:
-            # Try to find JSON in code blocks first
-            code_block_match = re.search(
-                r"```(?:json)?\s*(\{.*?\})\s*```", response, re.DOTALL
-            )
-            if code_block_match:
-                return json.loads(code_block_match.group(1))
-
-            # Try to find bare JSON
-            json_match = re.search(r"\{.*\}", response, re.DOTALL)
-            if json_match:
-                try:
-                    return json.loads(json_match.group())
-                except json.JSONDecodeError:
-                    # JSON might be truncated - extract what we can
-                    pass
-
-            # Fallback: extract scores using regex patterns for truncated JSON
-            result = default_result.copy()
-
-            quality_match = re.search(r'"quality_score"\s*:\s*([\d.]+)', response)
-            if quality_match:
-                result["quality_score"] = float(quality_match.group(1))
-
-            coherence_match = re.search(r'"coherence_score"\s*:\s*([\d.]+)', response)
-            if coherence_match:
-                result["coherence_score"] = float(coherence_match.group(1))
-
-            # Extract grammar issues array if present
-            grammar_match = re.search(
-                r'"grammar_issues"\s*:\s*\[(.*?)\]', response, re.DOTALL
-            )
-            if grammar_match:
-                try:
-                    issues_str = "[" + grammar_match.group(1) + "]"
-                    result["grammar_issues"] = json.loads(issues_str)
-                except json.JSONDecodeError:
-                    pass
-
-            return result
-        except (json.JSONDecodeError, ValueError):
-            return default_result
+__all__ = ["AIEnhancer"]
